Update i18n-js to version 4
