+++ app/javascript/packs/application.js
@@ -1,0 +1,3 @@
+require('./mirador_config.js')
+import { I18n } from 'i18n-js/dist/import/index.js';
+export const i18n = new I18n();
