+++ app/views/catalog/_mirador_default.html.erb
@@ -1,4 +1,2 @@
 <div id="m3" data-iiif-manifest="<%= document.iiif_manifest_url %>">
 </div>
-
-<%= javascript_pack_tag 'mirador' %>
