+++ app/views/layouts/spotlight/base.html.erb
@@ -19,6 +19,7 @@
     <%= twitter_card %>
     <%= opengraph %>
     <%= javascript_importmap_tags 'entry' %>
+    <%= javascript_pack_tag 'application', defer: true %>
     <%= javascript_include_tag "application", defer: true %>
     <%= javascript_tag "window.addEventListener('load', () => Spotlight.sirTrevorIcon = '#{asset_path('spotlight/blocks/sir-trevor-icons.svg')}')" %>
     <%= javascript_tag "window.addEventListener('load', () => $.fx.off = true)" if Rails.env.test? %>
