+++ app/views/shared/_i18n_js.html.erb
@@ -1,6 +1,6 @@
 <%= javascript_tag do %>
   window.addEventListener('load', () => {
-    I18n.defaultLocale = "<%= I18n.default_locale %>";
-    I18n.locale = "<%= I18n.locale %>";
+    i18n.defaultLocale = "<%= I18n.default_locale %>";
+    i18n.locale = "<%= I18n.locale %>";
   })
 <% end %>
