+++ package.json
@@ -17,6 +17,7 @@
     "bootstrap": "^4",
     "browser": "^0.2.6",
     "compression-webpack-plugin": "9",
+    "i18n-js": "^4.2.3",
     "mirador": "^3.3.0",
     "mirador-image-tools": "^0.11.0",
     "react": "^16.8.3",
